feat: almost done

+++ TodoList/Views/ListView.swift
@@ -12,15 +12,22 @@ struct ListView: View {
     @EnvironmentObject var listViewModel: ListViewModel
     
     var body: some View {
-        List {
-            ForEach(listViewModel.items) { item in
-                ListRowView(item: item)
+        ZStack { 
+            if listViewModel.items.isEmpty {
+                NoItemView()
+                    .transition(AnyTransition.opacity.animation(.easeIn))
+            } else {
+                List {
+                    ForEach(listViewModel.items) { item in
+                        ListRowView(item: item)
+                    }
+                    .onMove(perform: listViewModel.moveItem)
+                    .onDelete(perform: listViewModel.deleteItem)
+                    
+                }
+                .listStyle(.plain)
             }
-            .onMove(perform: listViewModel.moveItem)
-            .onDelete(perform: listViewModel.deleteItem)
-            
         }
-        .listStyle(.plain)
         .navigationTitle("Todo List")
         .navigationBarItems(
             leading: EditButton(),
@@ -37,4 +44,5 @@ struct ListView: View {
         ListView()
     }
     .environmentObject(ListViewModel())
+    //.preferredColorScheme(.dark) no work
 }
